batch_requirements.txt
# Batch DICOM Segmentation Requirements
# Install with: pip install -r batch_requirements.txt

# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pillow>=8.0.0
matplotlib>=3.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scipy>=1.7.0

# DICOM handling
pydicom>=2.3.0

# SAM (Segment Anything Model)
# Note: Install from GitHub as it's not on PyPI
# git+https://github.com/facebookresearch/segment-anything.git

# Optional but recommended for better performance
# timm>=0.6.0  # For vision transformers